+++ include/dsd-neo/runtime/input_ring.h
@@ -5,7 +5,7 @@
 
 /**
  * @file
- * @brief Input ring buffer API for interleaved I/Q int16_t samples.
+ * @brief Input ring buffer API for interleaved I/Q float samples.
  *
  * Declares the simple SPSC input ring and operations to reserve, commit,
  * write, and blockingly read samples with wrap-around handling.
@@ -18,10 +18,10 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-/* Simple SPSC ring for interleaved I/Q int16_t samples (input path) */
+/* Simple SPSC ring for interleaved I/Q float samples (input path) */
 struct input_ring_state {
-    int16_t* buffer;
-    size_t capacity; /* in int16_t elements */
+    float* buffer;
+    size_t capacity; /* in float elements */
     std::atomic<size_t> head;
     std::atomic<size_t> tail;
     pthread_cond_t ready;
@@ -79,8 +79,7 @@ input_ring_clear(struct input_ring_state* r) {
  * @param n2         [out] Second writable region length.
  * @return Total writable samples granted across regions.
  */
-int input_ring_reserve(struct input_ring_state* r, size_t min_needed, int16_t** p1, size_t* n1, int16_t** p2,
-                       size_t* n2);
+int input_ring_reserve(struct input_ring_state* r, size_t min_needed, float** p1, size_t* n1, float** p2, size_t* n2);
 
 /**
  * @brief Commit previously reserved writable regions to the input ring.
@@ -97,7 +96,7 @@ void input_ring_commit(struct input_ring_state* r, size_t produced);
  * @param data  Source samples to write.
  * @param count Number of samples to write.
  */
-void input_ring_write(struct input_ring_state* r, const int16_t* data, size_t count);
+void input_ring_write(struct input_ring_state* r, const float* data, size_t count);
 
 /**
  * @brief Read up to max_count samples from the input ring, blocking until available.
@@ -107,4 +106,4 @@ void input_ring_write(struct input_ring_state* r, const int16_t* data, size_t co
  * @param max_count Maximum number of samples to read.
  * @return Number of samples read (>=1), 0 if max_count is 0, or -1 on exit.
  */
-int input_ring_read_block(struct input_ring_state* r, int16_t* out, size_t max_count);
+int input_ring_read_block(struct input_ring_state* r, float* out, size_t max_count);
