+++ src/runtime/ring.cpp
@@ -27,7 +27,7 @@ extern volatile uint8_t exitflag; // defined in apps/dsd-cli/main.c
  * @param count Number of samples to write.
  */
 void
-ring_write(struct output_state* o, const int16_t* data, size_t count) {
+ring_write(struct output_state* o, const float* data, size_t count) {
     if (!o || !o->buffer) {
         return;
     }
@@ -65,7 +65,7 @@ ring_write(struct output_state* o, const int16_t* data, size_t count) {
         /* First region: from head to end of buffer */
         size_t to_end = o->capacity - h;
         if (to_end >= write_now) {
-            memcpy(o->buffer + h, data, write_now * sizeof(int16_t));
+            memcpy(o->buffer + h, data, write_now * sizeof(float));
             h += write_now;
             if (h >= o->capacity) {
                 h = 0;
@@ -78,12 +78,12 @@ ring_write(struct output_state* o, const int16_t* data, size_t count) {
 
         /* Handle wrap-around case: split write_now across tail and head */
         if (to_end > 0) {
-            memcpy(o->buffer + h, data, to_end * sizeof(int16_t));
+            memcpy(o->buffer + h, data, to_end * sizeof(float));
             data += to_end;
         }
         size_t remaining = write_now - to_end;
         if (remaining > 0) {
-            memcpy(o->buffer, data, remaining * sizeof(int16_t));
+            memcpy(o->buffer, data, remaining * sizeof(float));
             h = remaining;
             data += remaining;
         } else {
@@ -109,7 +109,7 @@ ring_write(struct output_state* o, const int16_t* data, size_t count) {
  * @param count Number of samples to write.
  */
 void
-ring_write_no_signal(struct output_state* o, const int16_t* data, size_t count) {
+ring_write_no_signal(struct output_state* o, const float* data, size_t count) {
     if (!o || !o->buffer) {
         return;
     }
@@ -146,7 +146,7 @@ ring_write_no_signal(struct output_state* o, const int16_t* data, size_t count) 
         /* First region: from head to end of buffer */
         size_t to_end = o->capacity - h;
         if (to_end >= write_now) {
-            memcpy(o->buffer + h, data, write_now * sizeof(int16_t));
+            memcpy(o->buffer + h, data, write_now * sizeof(float));
             h += write_now;
             if (h >= o->capacity) {
                 h = 0;
@@ -159,12 +159,12 @@ ring_write_no_signal(struct output_state* o, const int16_t* data, size_t count) 
 
         /* Handle wrap-around case: split write_now across tail and head */
         if (to_end > 0) {
-            memcpy(o->buffer + h, data, to_end * sizeof(int16_t));
+            memcpy(o->buffer + h, data, to_end * sizeof(float));
             data += to_end;
         }
         size_t remaining = write_now - to_end;
         if (remaining > 0) {
-            memcpy(o->buffer, data, remaining * sizeof(int16_t));
+            memcpy(o->buffer, data, remaining * sizeof(float));
             h = remaining;
             data += remaining;
         } else {
@@ -183,7 +183,7 @@ ring_write_no_signal(struct output_state* o, const int16_t* data, size_t count) 
  * @param count Number of samples to write.
  */
 void
-ring_write_signal_on_empty_transition(struct output_state* o, const int16_t* data, size_t count) {
+ring_write_signal_on_empty_transition(struct output_state* o, const float* data, size_t count) {
     int need_signal = ring_is_empty(o);
     ring_write_no_signal(o, data, count);
     if (need_signal) {
@@ -201,7 +201,7 @@ ring_write_signal_on_empty_transition(struct output_state* o, const int16_t* dat
  * @return 0 on success, -1 on exit.
  */
 int
-ring_read_one(struct output_state* o, int16_t* out) {
+ring_read_one(struct output_state* o, float* out) {
     if (!o || !o->buffer) {
         return -1;
     }
@@ -258,7 +258,7 @@ ring_read_one(struct output_state* o, int16_t* out) {
  * @return Number of samples read (>=1) or -1 on exit.
  */
 int
-ring_read_batch(struct output_state* o, int16_t* out, size_t max_count) {
+ring_read_batch(struct output_state* o, float* out, size_t max_count) {
     if (max_count == 0) {
         return 0;
     }
@@ -298,14 +298,14 @@ ring_read_batch(struct output_state* o, int16_t* out, size_t max_count) {
     size_t t = o->tail.load();
     size_t first = o->capacity - t;
     if (first >= read_now) {
-        memcpy(out, o->buffer + t, read_now * sizeof(int16_t));
+        memcpy(out, o->buffer + t, read_now * sizeof(float));
         t += read_now;
         if (t >= o->capacity) {
             t = 0;
         }
     } else {
-        memcpy(out, o->buffer + t, first * sizeof(int16_t));
-        memcpy(out + first, o->buffer, (read_now - first) * sizeof(int16_t));
+        memcpy(out, o->buffer + t, first * sizeof(float));
+        memcpy(out + first, o->buffer, (read_now - first) * sizeof(float));
         t = read_now - first;
     }
     o->tail.store(t);
